(github #21): Revive 32 bit support.
Scanning was working correctly but reporting and other logging was
broken on 32 bit builds. Do one pass sanitizing types, mostly getting
rid of long and off_t use. With this, 32 bit appears to be working
(tested on FreeBSD) and it also cleans up assorted warnings elsewhere.

+++ src/hashlist.h
@@ -115,7 +115,7 @@ void add_to_hash_table(struct hash_table * hl,
  *
  */
 void add_hash_table(struct hash_table * hl, struct path_list_entry * file,
-                    uint64_t blocks, int bsize, off_t skip);
+                    uint64_t blocks, int bsize, uint64_t skip);
 
 
 /** ***************************************************************************
@@ -134,7 +134,7 @@ void add_hash_table(struct hash_table * hl, struct path_list_entry * file,
  */
 void add_hash_table_from_mem(struct hash_table * hl,
                              struct path_list_entry * file,
-                             const char * buffer, off_t bufsize);
+                             const char * buffer, uint32_t bufsize);
 
 
 /** ***************************************************************************
@@ -156,7 +156,7 @@ void add_hash_table_from_mem(struct hash_table * hl,
  *
  */
 void publish_duplicate_hash_table(sqlite3 * dbh, struct hash_table * hl,
-                                  off_t size, int round);
+                                  uint64_t size, int round);
 
 
 /** ***************************************************************************
